ref: implement "input_size" argument for memory measure binaries
