--- binius/src/bin/sha256_2048_binius_with_lookup_mem.rs
@@ -1,9 +1,0 @@
-use binius::bench::{prove, sha256_with_lookup_prepare};
-
-fn main() {
-    let allocator = bumpalo::Bump::new();
-
-    let (constraint_system, args, witness, backend) = sha256_with_lookup_prepare(&allocator);
-
-    let (_, _, _proof) = prove(constraint_system, args, witness, backend);
-}
