--- plonky2/src/bin/sha256_2048_plonky2_no_lookup_mem.rs
@@ -1,6 +1,0 @@
-use plonky2_sha256::bench::{prove, sha256_no_lookup_prepare};
-
-fn main() {
-    let (data, pw) = sha256_no_lookup_prepare();
-    let _proof = prove(&data.prover_data(), pw);
-}
