+++ plonky3-powdr/Cargo.toml
@@ -28,6 +28,7 @@ rand = { workspace = true }
 sha2 = { workspace = true }
 criterion = { workspace = true }
 utils = { workspace = true }
+clap = { workspace = true }
 
 [[bench]]
 name = "prove_verify"
@@ -38,8 +39,8 @@ name = "measure_mem"
 path = "src/bin/measure_mem.rs"
 
 [[bin]]
-name = "sha256_2048_powdr_mem"
-path = "src/bin/sha256_2048_powdr_mem.rs"
+name = "sha256_mem"
+path = "src/bin/sha256_mem.rs"
 
 [[bin]]
 name = "plonky3-powdr-sha"
