--- plonky3-powdr/src/bin/sha256_2048_powdr_mem.rs
@@ -1,7 +1,0 @@
-use sha::bench::{prepare_pipeline, prove};
-
-fn main() {
-    let mut pipeline = prepare_pipeline();
-
-    prove(&mut pipeline);
-}
