+++ provekit/src/bin/measure_mem.rs
@@ -3,9 +3,9 @@ use utils::metadata::SHA2_INPUTS;
 
 fn main() {
     let script = "../measure_mem_avg.sh";
+    let binary_name = "sha256_mem";
     for input_size in SHA2_INPUTS {
         let json_file = format!("sha256_{}_provekit_mem_report.json", input_size);
-        let binary_name = format!("sha256_{}_provekit_mem", input_size);
         let binary_path = format!("../target/release/{}", binary_name);
         let _compile_output = Command::new("cargo")
             .arg("build")
@@ -21,6 +21,8 @@ fn main() {
             .arg(json_file)
             .arg("--")
             .arg(binary_path)
+            .arg("--input-size")
+            .arg(input_size.to_string())
             .output()
             .expect("failed to execute script");
