--- provekit/src/bin/sha256_2048_provekit_mem.rs
@@ -1,6 +1,0 @@
-use provekit::ProvekitSha256Benchmark;
-
-fn main() {
-    let bench_harness = ProvekitSha256Benchmark::new(2048);
-    let _proof = bench_harness.run_prove();
-}
